#143 - KB search renaming and other changes
