--- sti-kbsearch/src/main/java/uk/ac/shef/dcs/kbsearch/model/Clazz.java
@@ -1,27 +1,0 @@
-package uk.ac.shef.dcs.kbsearch.model;
-
-import java.util.Set;
-
-/**
- * Created by - on 15/03/2016.
- */
-public class Clazz extends Resource{
-    private static final long serialVersionUID = -1208425814000474692L;
-
-    public Clazz (String id, String label){
-        this.id=id;
-        this.label=label;
-    }
-
-    public String toString(){
-        return "id="+id+", label="+label;
-    }
-
-    public boolean equals(Object o){
-        if (o instanceof Clazz){
-            Clazz c = (Clazz)o;
-            return c.getId().equals(id) && c.getLabel().equals(label);
-        }
-        return false;
-    }
-}
